# api-server-mcp-demo/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: support@example.com
    name: API Support
  description: API Server for e-commerce product management system
  title: Product Management API
  version: '2.0'
servers:
- description: Development server
  url: http://localhost:8000/v1
- description: Production server
  url: https://api.example.com/v1
- description: Staging server
  url: https://staging-api.example.com/v1
paths:
...
  /orders/user/{user_id}:
    get:
      parameters:
      - description: The user ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Filter by order status
        in: query
        name: status
        required: false
        type: string
      - description: Filter by date range
        enum:
        - last_week
        - last_month
        - last_year
        in: query
        name: date_range
        required: false
        type: string
      responses:
        '200':
          description: List of user orders
          schema:
            items:
              type: object
            type: array
        '400':
          description: Invalid parameters
        '404':
          description: User not found
      summary: Get all orders for a user by ID
...
  /users/:
    get:
      parameters:
      - description: Search users by name (case-insensitive partial match)
        in: query
        name: name
        required: false
        type: string
      - description: Filter users by email (case-insensitive exact match)
        in: query
        name: email
        required: false
        type: string
      responses:
        '200':
          description: List of users
          schema:
            items:
              properties:
                email:
                  type: string
                name:
                  type: string
                user_id:
                  type: string
              type: object
            type: array
        '404':
          description: No users found
      summary: Get all users or search by criteria
  /users/{user_id}:
    get:
      parameters:
      - description: The user ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        '200':
          description: User details
          schema:
            properties:
              email:
                type: string
              name:
                type: string
              user_id:
                type: string
            type: object
        '404':
          description: User not found
      summary: Get user by ID